# server/docs/swagger.yaml
definitions:
...
  config.AiCoureConfig:
    properties:
      url:
        type: string
    type: object
  config.AliyunOSS:
    properties:
      access-key-id:
        type: string
      access-key-secret:
        type: string
      base-path:
        type: string
      bucket-name:
        type: string
      bucket-url:
        type: string
      endpoint:
        type: string
    type: object
  config.Autocode:
    properties:
      ai-path:
        type: string
      module:
        type: string
      root:
        type: string
      server:
        type: string
      web:
        type: string
    type: object
  config.AwsS3:
    properties:
      base-url:
        type: string
      bucket:
        type: string
      disable-ssl:
        type: boolean
      endpoint:
        type: string
      path-prefix:
        type: string
      region:
        type: string
      s3-force-path-style:
        type: boolean
      secret-id:
        type: string
      secret-key:
        type: string
    type: object
  config.CORS:
    properties:
      mode:
        type: string
      whitelist:
        items:
          $ref: '#/definitions/config.CORSWhitelist'
        type: array
    type: object
  config.CORSWhitelist:
    properties:
      allow-credentials:
        type: boolean
      allow-headers:
        type: string
      allow-methods:
        type: string
      allow-origin:
        type: string
      expose-headers:
        type: string
    type: object
  config.Captcha:
    properties:
      img-height:
        description: 验证码高度
        type: integer
      img-width:
        description: 验证码宽度
        type: integer
      key-long:
        description: 验证码长度
        type: integer
      open-captcha:
        description: 防爆破验证码开启此数，0代表每次登录都需要验证码，其他数字代表错误密码此数，如3代表错误三次后出现验证码
        type: integer
      open-captcha-timeout:
        description: 防爆破验证码超时时间，单位：s(秒)
        type: integer
    type: object
  config.CloudflareR2:
    properties:
      access-key-id:
        type: string
      account-id:
        type: string
      base-url:
        type: string
      bucket:
        type: string
      path:
        type: string
      secret-access-key:
        type: string
    type: object
  config.DiskList:
    properties:
      mount-point:
        type: string
    type: object
  config.Excel:
    properties:
      dir:
        type: string
    type: object
  config.HuaWeiObs:
    properties:
      access-key:
        type: string
      bucket:
        type: string
      endpoint:
        type: string
      path:
        type: string
      secret-key:
        type: string
    type: object
  config.JWT:
    properties:
      buffer-time:
        description: 缓冲时间
        type: string
      expires-time:
        description: 过期时间
        type: string
      issuer:
        description: 签发者
        type: string
      signing-key:
        description: jwt签名
        type: string
    type: object
  config.Local:
    properties:
      path:
        description: 本地文件访问路径
        type: string
      store-path:
        description: 本地文件存储路径
        type: string
    type: object
  config.Minio:
    properties:
      access-key-id:
        type: string
      access-key-secret:
        type: string
      base-path:
        type: string
      bucket-name:
        type: string
      bucket-url:
        type: string
      endpoint:
        type: string
      use-ssl:
        type: boolean
    type: object
  config.Mongo:
    properties:
      auth-source:
        description: 验证数据库
        type: string
      coll:
        description: collection name
        type: string
      connect-timeout-ms:
        description: 连接超时时间
        type: integer
      database:
        description: database name
        type: string
      hosts:
        description: 主机列表
        items:
          $ref: '#/definitions/config.MongoHost'
        type: array
      is-zap:
        description: 是否开启zap日志
        type: boolean
      max-pool-size:
        description: 最大连接池
        type: integer
      min-pool-size:
        description: 最小连接池
        type: integer
      options:
        description: mongodb options
        type: string
      password:
        description: 密码
        type: string
      socket-timeout-ms:
        description: socket超时时间
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  config.MongoHost:
    properties:
      host:
        description: ip地址
        type: string
      port:
        description: 端口
        type: string
    type: object
  config.Mssql:
    properties:
      config:
        description: 高级配置
        type: string
      db-name:
        description: 数据库名
        type: string
      engine:
        default: InnoDB
        description: 数据库引擎，默认InnoDB
        type: string
      log-mode:
        description: 是否开启Gorm全局日志
        type: string
      log-zap:
        description: 是否通过zap写入日志文件
        type: boolean
      max-idle-conns:
        description: 空闲中的最大连接数
        type: integer
      max-open-conns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 数据库地址
        type: string
      port:
        description: 数据库端口
        type: string
      prefix:
        description: 数据库前缀
        type: string
      singular:
        description: 是否开启全局禁用复数，true表示开启
        type: boolean
      username:
        description: 数据库账号
        type: string
    type: object
  config.Mysql:
    properties:
      config:
        description: 高级配置
        type: string
      db-name:
        description: 数据库名
        type: string
      engine:
        default: InnoDB
        description: 数据库引擎，默认InnoDB
        type: string
      log-mode:
        description: 是否开启Gorm全局日志
        type: string
      log-zap:
        description: 是否通过zap写入日志文件
        type: boolean
      max-idle-conns:
        description: 空闲中的最大连接数
        type: integer
      max-open-conns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 数据库地址
        type: string
      port:
        description: 数据库端口
        type: string
      prefix:
        description: 数据库前缀
        type: string
      singular:
        description: 是否开启全局禁用复数，true表示开启
        type: boolean
      username:
        description: 数据库账号
        type: string
    type: object
  config.Oracle:
    properties:
      config:
        description: 高级配置
        type: string
      db-name:
        description: 数据库名
        type: string
      engine:
        default: InnoDB
        description: 数据库引擎，默认InnoDB
        type: string
      log-mode:
        description: 是否开启Gorm全局日志
        type: string
      log-zap:
        description: 是否通过zap写入日志文件
        type: boolean
      max-idle-conns:
        description: 空闲中的最大连接数
        type: integer
      max-open-conns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 数据库地址
        type: string
      port:
        description: 数据库端口
        type: string
      prefix:
        description: 数据库前缀
        type: string
      singular:
        description: 是否开启全局禁用复数，true表示开启
        type: boolean
      username:
        description: 数据库账号
        type: string
    type: object
  config.Pgsql:
    properties:
      config:
        description: 高级配置
        type: string
      db-name:
        description: 数据库名
        type: string
      engine:
        default: InnoDB
        description: 数据库引擎，默认InnoDB
        type: string
      log-mode:
        description: 是否开启Gorm全局日志
        type: string
      log-zap:
        description: 是否通过zap写入日志文件
        type: boolean
      max-idle-conns:
        description: 空闲中的最大连接数
        type: integer
      max-open-conns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 数据库地址
        type: string
      port:
        description: 数据库端口
        type: string
      prefix:
        description: 数据库前缀
        type: string
      singular:
        description: 是否开启全局禁用复数，true表示开启
        type: boolean
      username:
        description: 数据库账号
        type: string
    type: object
  config.Qiniu:
    properties:
      access-key:
        description: 秘钥AK
        type: string
      bucket:
        description: 空间名称
        type: string
      img-path:
        description: CDN加速域名
        type: string
      secret-key:
        description: 秘钥SK
        type: string
      use-cdn-domains:
        description: 上传是否使用CDN上传加速
        type: boolean
      use-https:
        description: 是否使用https
        type: boolean
      zone:
        description: 存储区域
        type: string
    type: object
  config.Redis:
    properties:
      addr:
        description: 服务器地址:端口
        type: string
      clusterAddrs:
        description: 集群模式下的节点地址列表
        items:
          type: string
        type: array
      db:
        description: 单实例模式下redis的哪个数据库
        type: integer
      name:
        description: 代表当前实例的名字
        type: string
      password:
        description: 密码
        type: string
      useCluster:
        description: 是否使用集群模式
        type: boolean
    type: object
  config.Server:
    properties:
      ai_core:
        allOf:
        - $ref: '#/definitions/config.AiCoureConfig'
        description: ai_core
      aliyun-oss:
        $ref: '#/definitions/config.AliyunOSS'
      autocode:
        allOf:
        - $ref: '#/definitions/config.Autocode'
        description: auto
      aws-s3:
        $ref: '#/definitions/config.AwsS3'
      captcha:
        $ref: '#/definitions/config.Captcha'
      cloudflare-r2:
        $ref: '#/definitions/config.CloudflareR2'
      cors:
        allOf:
        - $ref: '#/definitions/config.CORS'
        description: 跨域配置
      db-list:
        items:
          $ref: '#/definitions/config.SpecializedDB'
        type: array
      disk-list:
        items:
          $ref: '#/definitions/config.DiskList'
        type: array
      email:
        $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_config.Email'
      excel:
        $ref: '#/definitions/config.Excel'
      hua-wei-obs:
        $ref: '#/definitions/config.HuaWeiObs'
      jwt:
        $ref: '#/definitions/config.JWT'
      local:
        allOf:
        - $ref: '#/definitions/config.Local'
        description: oss
      minio:
        $ref: '#/definitions/config.Minio'
      mongo:
        $ref: '#/definitions/config.Mongo'
      mssql:
        $ref: '#/definitions/config.Mssql'
      mysql:
        allOf:
        - $ref: '#/definitions/config.Mysql'
        description: gorm
      oracle:
        $ref: '#/definitions/config.Oracle'
      pgsql:
        $ref: '#/definitions/config.Pgsql'
      qiniu:
        $ref: '#/definitions/config.Qiniu'
      redis:
        $ref: '#/definitions/config.Redis'
      redis-list:
        items:
          $ref: '#/definitions/config.Redis'
        type: array
      sqlite:
        $ref: '#/definitions/config.Sqlite'
      system:
        $ref: '#/definitions/config.System'
      tencent-cos:
        $ref: '#/definitions/config.TencentCOS'
      zap:
        $ref: '#/definitions/config.Zap'
    type: object
  config.SpecializedDB:
    properties:
      alias-name:
        type: string
      config:
        description: 高级配置
        type: string
      db-name:
        description: 数据库名
        type: string
      disable:
        type: boolean
      engine:
        default: InnoDB
        description: 数据库引擎，默认InnoDB
        type: string
      log-mode:
        description: 是否开启Gorm全局日志
        type: string
      log-zap:
        description: 是否通过zap写入日志文件
        type: boolean
      max-idle-conns:
        description: 空闲中的最大连接数
        type: integer
      max-open-conns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 数据库地址
        type: string
      port:
        description: 数据库端口
        type: string
      prefix:
        description: 数据库前缀
        type: string
      singular:
        description: 是否开启全局禁用复数，true表示开启
        type: boolean
      type:
        type: string
      username:
        description: 数据库账号
        type: string
    type: object
  config.Sqlite:
    properties:
      config:
        description: 高级配置
        type: string
      db-name:
        description: 数据库名
        type: string
      engine:
        default: InnoDB
        description: 数据库引擎，默认InnoDB
        type: string
      log-mode:
        description: 是否开启Gorm全局日志
        type: string
      log-zap:
        description: 是否通过zap写入日志文件
        type: boolean
      max-idle-conns:
        description: 空闲中的最大连接数
        type: integer
      max-open-conns:
        description: 打开到数据库的最大连接数
        type: integer
      password:
        description: 数据库密码
        type: string
      path:
        description: 数据库地址
        type: string
      port:
        description: 数据库端口
        type: string
      prefix:
        description: 数据库前缀
        type: string
      singular:
        description: 是否开启全局禁用复数，true表示开启
        type: boolean
      username:
        description: 数据库账号
        type: string
    type: object
  config.System:
    properties:
      addr:
        description: 端口值
        type: integer
      db-type:
        description: 数据库类型:mysql(默认)|sqlite|sqlserver|postgresql
        type: string
      iplimit-count:
        type: integer
      iplimit-time:
        type: integer
      oss-type:
        description: Oss类型
        type: string
      router-prefix:
        type: string
      use-mongo:
        description: 使用mongo
        type: boolean
      use-multipoint:
        description: 多点登录拦截
        type: boolean
      use-redis:
        description: 使用redis
        type: boolean
      use-strict-auth:
        description: 使用树形角色分配模式
        type: boolean
    type: object
  config.TencentCOS:
    properties:
      base-url:
        type: string
      bucket:
        type: string
      path-prefix:
        type: string
      region:
        type: string
      secret-id:
        type: string
      secret-key:
        type: string
    type: object
  config.Zap:
    properties:
      director:
        description: 日志文件夹
        type: string
      encode-level:
        description: 编码级
        type: string
      format:
        description: 输出
        type: string
      level:
        description: 级别
        type: string
      log-in-console:
        description: 输出控制台
        type: boolean
      prefix:
        description: 日志前缀
        type: string
      retention-day:
        description: 日志保留天数
        type: integer
      show-line:
        description: 显示行
        type: boolean
      stacktrace-key:
        description: 栈名
        type: string
    type: object
...
  github_com_flipped-aurora_gin-vue-admin_server_config.Email:
    properties:
      from:
        description: 发件人  你自己要发邮件的邮箱
        type: string
      host:
        description: 服务器地址 例如 smtp.qq.com  请前往QQ或者你要发邮件的邮箱查看其smtp协议
        type: string
      is-ssl:
        description: 是否SSL   是否开启SSL
        type: boolean
      nickname:
        description: 昵称    发件人昵称 通常为自己的邮箱
        type: string
      port:
        description: 端口     请前往QQ或者你要发邮件的邮箱查看其smtp协议 大多为 465
        type: integer
      secret:
        description: 密钥    用于登录的密钥 最好不要用邮箱密码 去邮箱smtp申请一个用于登录的密钥
        type: string
      to:
        description: 收件人:多个以英文逗号分隔 例：a@qq.com b@qq.com 正式开发中请把此项目作为参数使用
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.SysConfigResponse:
    properties:
      config:
        $ref: '#/definitions/config.Server'
    type: object
...
  system.SysParams:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      desc:
        description: 参数说明
        type: string
      key:
        description: 参数键
        type: string
      name:
        description: 参数名称
        type: string
      updatedAt:
        description: 更新时间
        type: string
      value:
        description: 参数值
        type: string
    required:
    - key
    - name
    - value
    type: object
...
  system.System:
    properties:
      config:
        $ref: '#/definitions/config.Server'
    type: object
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.7.7
paths:
...
  /sysParams/createSysParams:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysParams'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建参数
      tags:
      - SysParams
  /sysParams/deleteSysParams:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysParams'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除参数
      tags:
      - SysParams
  /sysParams/deleteSysParamsByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除参数
      tags:
      - SysParams
  /sysParams/findSysParams:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 参数说明
        in: query
        name: desc
        type: string
      - description: 参数键
        in: query
        name: key
        required: true
        type: string
      - description: 参数名称
        in: query
        name: name
        required: true
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 参数值
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.SysParams'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询参数
      tags:
      - SysParams
  /sysParams/getSysParam:
    get:
      consumes:
      - application/json
      parameters:
      - description: key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.SysParams'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 根据key获取参数value
      tags:
      - SysParams
  /sysParams/getSysParamsList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: key
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取参数列表
      tags:
      - SysParams
  /sysParams/updateSysParams:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysParams'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新参数
      tags:
      - SysParams
  /system/getServerInfo:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 获取服务器信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取服务器信息
      tags:
      - System
  /system/getSystemConfig:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 获取配置文件内容,返回包括系统配置
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.SysConfigResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取配置文件内容
      tags:
      - System
  /system/reloadSystem:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 重启系统
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 重启系统
      tags:
      - System
  /system/setSystemConfig:
    post:
      parameters:
      - description: 设置配置文件内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.System'
      produces:
      - application/json
      responses:
        "200":
          description: 设置配置文件内容
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置配置文件内容
      tags:
      - System
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"